refactor: split the hot-reload orchestrator below the file-length limit
The 500-line gate cannot land while this 2899-SLOC host stays one file.
Extract cohesive helper groups and move three stage methods whole onto
their owners so ProcessFileAsync stays a thin driver under the limit.
