+++ Packages/src/Editor/FirstPartyTools/HotReload/HotReloadAppliedSourceLifecycle.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 420508d95eeac49b09803f7776666e98
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
