+++ Packages/src/Editor/FirstPartyTools/HotReload/HotReloadEntryApplier.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 0fdf06c3daea94ae2af1bd5aed5fced3
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
