+++ Packages/src/Editor/FirstPartyTools/HotReload/HotReloadOrchestratorLog.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5b960831995674cbfa86133352c77ae2
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
