+++ Packages/src/Editor/FirstPartyTools/HotReload/HotReloadOutcomeAggregation.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ae72797b26bb942ceb073679099091ca
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
