+++ Packages/src/Editor/FirstPartyTools/HotReload/HotReloadPatchTargetSupport.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: eb3fba1df10734fb39f56d606af9a743
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
