+++ Packages/src/Editor/FirstPartyTools/HotReload/HotReloadRemovedMembersWarning.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 722324f36b3834dce8327e565983decc
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
