+++ Packages/src/Editor/FirstPartyTools/HotReload/HotReloadShimErrorAttribution.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e8a42de5df1db4194b62cdf36d06da9c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
