+++ Packages/src/Editor/FirstPartyTools/HotReload/HotReloadShimFirstCompile.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 50b0cfcade27244309bf782d00c7e376
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
