+++ Packages/src/Editor/FirstPartyTools/HotReload/HotReloadShimIsolation.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 492827e2fce884acd98300e9e09407e7
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
