+++ Packages/src/Editor/FirstPartyTools/HotReload/HotReloadSignatureChangeCoverage.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 0c768fee3c1464866834262f8a32b3c4
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
