+++ Packages/src/Editor/FirstPartyTools/HotReload/HotReloadSignatureChangeGate.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ea72a6198f8324559acba423e85dbc86
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
