+++ Packages/src/Editor/FirstPartyTools/HotReload/HotReloadWireMethodKeys.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 299477112105548f59a0044400d69dbd
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
